Made a front end for debt settlement page, it shows open debts in current event. Added method to initialise page in MainCtrl.java, as well as rewrote a controller class for the page that populates the page with open debts in the current event, tested it on my computer with debugger it works; however in real life scenario it would not, because debt attribute is calculated for a person when new Expense is created and Transaction is initialised, but debt attribute is not updated in Data Base, we should fix it, I already notified creator of the AddExpenseCtrl in whatsup.

+++ commons/src/main/java/commons/DebtCellData.java
@@ -4,7 +4,7 @@ public class DebtCellData {
 
   protected Person sender;
   protected Person receiver;
-  protected int debt;
+  protected double debt;
 
   /**
    * constructor which initializes all attributes.
@@ -12,7 +12,7 @@ public class DebtCellData {
    * @param receiver
    * @param debt
    */
-  public DebtCellData(Person sender, Person receiver, int debt){
+  public DebtCellData(Person sender, Person receiver, double debt){
     this.sender = sender;
     this.receiver = receiver;
     this.debt = debt;
@@ -22,7 +22,7 @@ public class DebtCellData {
    * Getter for the amount of debt.
    * @return
    */
-  public int getDebt() {
+  public double getDebt() {
     return debt;
   }
